% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summaries.R
\name{get_summaries}
\alias{get_summaries}
\title{Get refseq gene summaries for provided gene identifiers}
\usage{
get_summaries(gene_ids)
}
\arguments{
\item{gene_ids}{A vector of gene identifiers. Entrez, ensembl, and gene symbols all supported.}
}
\value{
A character vector of gene summaries the same length as \code{gene_ids}.
}
\description{
This function uses the \code{\link[mygene:queryMany]{mygene::queryMany()}} to retrieve gene summaries from
any number of provided gene identifiers using the mygene API.
}
\examples{
genes <- c(1017, 1018, "ENSG00000148795", "CDK2", "goobledegook")
get_summaries(genes)
}
\seealso{
\code{\link[mygene:queryMany]{mygene::queryMany()}}
}
\author{
Jared Andrews
}
